refactor : Refactor BoPage
