+++ src/pages/BOPage/_shared/SearchInput/searchInput.module.scss
@@ -1,0 +1,40 @@
+@use "../../../../styles/constants/colors" as c;
+@use "../../../../styles/mixins" as m;
+
+.searchForm {
+  @include m.flex(flex-start);
+  width: 600px;
+  height: 50px;
+  border: none;
+
+  .searchInput {
+    width: calc(100% - 50px);
+    height: 100%;
+    padding-left: 20px;
+    border: none;
+    border-radius: 10px 0 0 10px;
+    background-color: c.$C_WHITE_40;
+    font-size: 18px;
+    color: c.$C_WHITE;
+
+    &::placeholder {
+      font-size: 18px;
+      color: c.$C_WHITE_70;
+    }
+  }
+
+  .searchBtn {
+    @include m.flex();
+    width: 50px;
+    height: 50px;
+    border: 0;
+    border-radius: 0 10px 10px 0;
+    background-color: c.$C_WHITE_40;
+    cursor: pointer;
+
+    .iconSearch {
+      width: 30px;
+      cursor: pointer;
+    }
+  }
+}
