+++ src/pages/BOPage/_shared/index.js
@@ -2,3 +2,4 @@ export { default as BoMovieModal } from "./BoMovieModal";
 export { default as BoUserModal } from "./BoUserModal";
 export { default as BoCommentModal } from "./BoCommentModal";
 export { default as BoReportModal } from "./BoReportModal";
+export { default as SearchInput } from "./SearchInput";
